rename credentials to external_credentials
